requirement.txt
streamlit
pdfplumber
PyPDF2
google-genai
pandas
beautifulsoup4
requests
python-dotenv
